requirements.txt
PyQt5>=5.15.0
python-vlc>=3.0.0
openai-whisper>=20231117
googletrans>=3.1.0a0
google-generativeai>=0.3.0
nltk>=3.8.1
Pillow>=10.0.0
pyinstaller>=6.0.0 